pyproject.toml
[tool.poetry]
name = "astro-gpt"
version = "0.1.0"
description = ""
authors = ["airtucha <airtucha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.5"
python-telegram-bot = "^21.0.1"
langchain = "^0.1.13"
langchain-core = "^0.1.33"
langchain-community = "^0.0.29"
langchain-openai = "^0.1.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
